sb/intel/i82371eb/fadt: Fix compilation on x86_64
Reviewed-on: https://review.coreboot.org/c/31471
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/southbridge/intel/i82371eb/fadt.c
@@ -46,8 +46,8 @@ void acpi_create_fadt(acpi_fadt_t *fadt, acpi_facs_t *facs, void *dsdt)
 	memcpy(header->asl_compiler_id, ASLC, 4);
 	header->asl_compiler_revision = 42;
 
-	fadt->firmware_ctrl = (u32)facs;
-	fadt->dsdt = (u32)dsdt;
+	fadt->firmware_ctrl = (uintptr_t)facs;
+	fadt->dsdt = (uintptr_t)dsdt;
 	fadt->preferred_pm_profile = 0; /* unspecified */
 	fadt->sci_int = 9;
 	fadt->smi_cmd = 0; /* smi command port */
@@ -153,9 +153,9 @@ void acpi_create_fadt(acpi_fadt_t *fadt, acpi_facs_t *facs, void *dsdt)
 	fadt->reset_reg.addrh = 0x0;
 	fadt->reset_value = 0;
 
-	fadt->x_firmware_ctl_l = (u32)facs;
+	fadt->x_firmware_ctl_l = (uintptr_t)facs;
 	fadt->x_firmware_ctl_h = 0;
-	fadt->x_dsdt_l = (u32)dsdt;
+	fadt->x_dsdt_l = (uintptr_t)dsdt;
 	fadt->x_dsdt_h = 0;
 
 	fadt->x_pm1a_evt_blk.space_id = 1;
